rename: lookupType typings -> 페이지 내 types 폴더로 이동
