+++ src/pages/lookup/types/lookupType.ts
